Renamed example folders for synapse rest and synapse graphql and added a simple postman collection
