+++ service/example-service-book-rest/src/main/java/io/americanexpress/service/book/rest/config/BookConfig.java
@@ -13,7 +13,6 @@
  */
 package io.americanexpress.service.book.rest.config;
 
-
 import io.americanexpress.synapse.service.rest.config.ServiceRestConfig;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
