+++ service/example-service-book-rest/src/main/java/io/americanexpress/service/book/rest/model/ReadBookResponse.java
@@ -14,8 +14,6 @@
 package io.americanexpress.service.book.rest.model;
 
 import io.americanexpress.synapse.service.rest.model.BaseServiceResponse;
-import io.swagger.v3.oas.annotations.media.Schema;
-
 import java.util.Objects;
 
 public class ReadBookResponse extends BaseServiceResponse {
