+++ service/pom.xml
@@ -27,8 +27,8 @@
 
     <!--Modules-->
     <modules>
-        <module>example-service-graphql-book</module>
-        <module>example-service-rest-book</module>
+        <module>example-service-book-graphql</module>
+        <module>example-service-book-rest</module>
         <module>synapse-service-graphql</module>
         <module>synapse-service-rest</module>
         <module>synapse-service-test</module>
